BEN-3407: Upgrade Filer to Django 5.1

+++ filer/models/filemodels.py
@@ -1,32 +1,28 @@
 #-*- coding: utf-8 -*-
-import polymorphic
 import hashlib
 import os
 import filer
 import logging
-import operator
 
 from django.contrib.auth import models as auth_models
 from django.urls import reverse
 from django.core.files.base import ContentFile
 from django.core.exceptions import ValidationError
 from django.db import (models, IntegrityError, transaction)
+from django.db.models import DEFERRED
 from django.utils.translation import gettext_lazy as _
 from filer.fields.multistorage_file import MultiStorageFileField
 from filer.models import mixins
 from filer.utils.cms_roles import *
 from filer.utils.files import matching_file_subtypes
 from filer import settings as filer_settings
+from filer.utils.cache import invalidate_folder_listing_cache, invalidate_folder_listing_cache_for_file
 from django.db.models import Count
 from django.utils import timezone
 
 from polymorphic.models import PolymorphicModel
 from polymorphic.managers import PolymorphicManager
 from polymorphic.query import PolymorphicQuerySet
-import hashlib
-import os
-import filer
-import logging
 
 logger = logging.getLogger(__name__)
 
@@ -91,7 +87,6 @@ class AliveFileManager(FileManager):
     # this is required in order to make sure that other models that are
     #   related to filer files will get an DoesNotExist exception if the file
     #   is in trash
-    use_for_related_fields = True
 
     def get_queryset(self):
         return super(AliveFileManager, self).get_queryset().filter(
@@ -159,6 +154,7 @@ class File(PolymorphicModel,
                     'snippet but will not be able to delete or '
                     'modify the current version of the asset.'))
 
+
     objects = AliveFileManager()
     trash = TrashFileManager()
     all_objects = FileManager()
@@ -167,15 +163,42 @@ class File(PolymorphicModel,
     def matches_file_type(cls, iname, ifile, request):
         return True  # I match all files...
 
+    # Sentinel for fields whose previous value is unknown (deferred).
+    # Using a dedicated sentinel instead of substituting empty defaults
+    # prevents false positives in change-detection that could trigger
+    # unnecessary file copies/moves on storage.
+    _UNKNOWN = object()
+
     def __init__(self, *args, **kwargs):
         super(File, self).__init__(*args, **kwargs)
-        self._old_is_public = self.is_public
-        self._old_sha1 = self.sha1
+        # Use __dict__ to avoid triggering deferred field loading
+        # which can cause recursion in Django 5.1+ (from_db calls __init__).
+        raw_is_public = self.__dict__.get('is_public', DEFERRED)
+        if raw_is_public is DEFERRED:
+            if self.pk is not None:
+                self._old_is_public = self.__class__.all_objects.filter(
+                    pk=self.pk
+                ).values_list('is_public', flat=True).first()
+            else:
+                self._old_is_public = False
+        else:
+            self._old_is_public = raw_is_public
+        raw_sha1 = self.__dict__.get('sha1', DEFERRED)
+        self._old_sha1 = self._UNKNOWN if raw_sha1 is DEFERRED else raw_sha1
         self._force_commit = False
         # see method _is_path_changed
-        self._old_name = self.name
-        self._current_file_location = self.file.name
-        self._old_folder_id = self.folder_id
+        raw_name = self.__dict__.get('name', DEFERRED)
+        self._old_name = self._UNKNOWN if raw_name is DEFERRED else raw_name
+        # For FileField, the raw value in __dict__ is the file name string
+        file_val = self.__dict__.get('file', '')
+        if file_val is DEFERRED:
+            file_val = ''
+        if file_val and hasattr(file_val, 'name'):
+            self._current_file_location = file_val.name
+        else:
+            self._current_file_location = file_val or ''
+        raw_folder_id = self.__dict__.get('folder_id', DEFERRED)
+        self._old_folder_id = self._UNKNOWN if raw_folder_id is DEFERRED else raw_folder_id
 
     def clean(self):
         if self.name:
@@ -237,7 +260,7 @@ class File(PolymorphicModel,
         src_file = src_storage.open(src_file_name)
         src_file.open()
         self.file = dst_storage.save(dst_file_name,
-            ContentFile(src_file.read()))
+            ContentFile(src_file.read(), name=os.path.basename(dst_file_name)))
         src_file.close()
         src_storage.delete(src_file_name)
 
@@ -269,7 +292,7 @@ class File(PolymorphicModel,
             if storage.exists(destination):
                 storage.delete(destination)
             destination = storage.save(destination,
-                                       ContentFile(file_content))
+                                       ContentFile(file_content, name=os.path.basename(destination)))
         self._current_file_location = destination
         self._old_name = self.name
         self._old_folder_id = getattr(self.folder, 'id', None)
@@ -314,7 +337,10 @@ class File(PolymorphicModel,
             self.generate_sha1()
         except (IOError, TypeError, ValueError) as e:
             pass
-        replaced_file = self._old_sha1 != self.sha1
+        replaced_file = (self._old_sha1 is not self._UNKNOWN and
+                         self._old_sha1 != self.sha1)
+        # Track old folder for cache invalidation when file moves between folders
+        old_folder_id = self._old_folder_id
         if filer_settings.FOLDER_AFFECTS_URL and (self._is_path_changed() or replaced_file):
             if replaced_file and not self._is_name_changed():
                 self.name = None  # if new file submitted for same id we overwrite what was previously in name
@@ -322,11 +348,25 @@ class File(PolymorphicModel,
             self.update_location_on_storage(*args, **kwargs)
         else:
             super(File, self).save(*args, **kwargs)
+        # Invalidate cache for the current folder
+        invalidate_folder_listing_cache_for_file(self)
+        # If file moved between folders, also invalidate the old folder
+        new_folder_id = getattr(self.folder, 'id', None)
+        if (old_folder_id is not self._UNKNOWN and
+                old_folder_id and old_folder_id != new_folder_id):
+            try:
+                old_folder = filer.models.foldermodels.Folder.all_objects.get(
+                    id=old_folder_id)
+                invalidate_folder_listing_cache(old_folder)
+            except filer.models.foldermodels.Folder.DoesNotExist:
+                pass
 
     save.alters_data = True
 
     def _is_name_changed(self):
         """Check if the file name was explicitly changed by the user."""
+        if self._old_name is self._UNKNOWN:
+            return False  # can't determine change from deferred field
         if self._old_name in ('', None):
             return self.name not in ('', None)
         return self._old_name != self.name
@@ -338,6 +378,11 @@ class File(PolymorphicModel,
             the values will be reset after the file is copied in the
             destination location on storage.
         """
+        # If previous values are unknown (deferred), skip change-detection
+        # to avoid triggering unnecessary file copies/moves on storage.
+        if self._old_name is self._UNKNOWN or self._old_folder_id is self._UNKNOWN:
+            return False
+
         # check if file name changed
         if self._old_name in ('', None):
             name_changed = self.name not in ('', None)
@@ -363,7 +408,14 @@ class File(PolymorphicModel,
         if self._old_sha1 != self.sha1:
             # actual file content needs to be replaced on storage prior to
             #   filer file instance save
-            self.file.storage.save(self._current_file_location, self.file)
+            if self._current_file_location:
+                self.file.storage.save(self._current_file_location, self.file)
+            else:
+                # New file — save to the computed target location directly
+                target = self.file.field.upload_to(self, self.upload_to_name)
+                saved_target = self.file.storage.save(target, self.file)
+                self._current_file_location = saved_target
+                self.file.name = saved_target
             self._old_sha1 = self.sha1
         new_location = self.file.field.upload_to(self, self.upload_to_name)
         storage = self.file.storage
@@ -388,12 +440,12 @@ class File(PolymorphicModel,
                     copy_and_save()
             except:
                 # delete the file from new_location if the db update failed
-                if old_location != new_location:
+                if old_location and old_location != new_location:
                     storage.delete(new_location)
                 raise
             else:
                 # only delete the file on the old_location if all went OK
-                if old_location != new_location:
+                if old_location and old_location != new_location:
                     storage.delete(old_location)
         else:
             copy_and_save()
@@ -444,6 +496,8 @@ class File(PolymorphicModel,
 
             self.deleted_at = deletion_time
             self.file = new_location
+        # Invalidate cache for the folder this file was in
+        invalidate_folder_listing_cache_for_file(self)
 
     def hard_delete(self, *args, **kwargs):
         """
@@ -533,6 +587,8 @@ class File(PolymorphicModel,
                     clipboard.append_file(File.objects.get(id=self.id))
                 except auth_models.User.DoesNotExist:
                     pass
+        # Invalidate cache for the folder this file was restored to
+        invalidate_folder_listing_cache_for_file(self)
 
     @property
     def label(self):
@@ -603,7 +659,7 @@ class File(PolymorphicModel,
         full_path = '{}{}{}'.format(directory_path, os.sep, self.actual_name)
         return full_path
 
-    def __unicode__(self):
+    def __str__(self):
         try:
             name = self.pretty_logical_path
         except:
@@ -759,8 +815,6 @@ class File(PolymorphicModel,
             return can_delete_file and has_role_on_site(user, self.folder.site)
         return False
 
-    def __str__(self):
-        return self.__unicode__()
     class Meta:
         app_label = 'filer'
         verbose_name = _('file')
